Enable AddressSanitizer and UndefinedBehaviourSanitizer
When building with DEBUG=ON, we automatically enable ASan and UBSan
to detect possible addressability issues, memory leaks, and dubious
code in the library.

Arguably, these sanitizers are orthogonal to debug support but we
enable these together for the time being so that we have fewer
possible configurations to test on GitHub actions.