fix simple_locked_queue implementation
without implementing the open/closed semantics, there was a race
condition, which allowed operations to be stored in the queue when there
was not actually any helper assigned to take care of it, leading to the
operation potentially never being executed.

This commit also re-adds the queue to the set of tested implementations.