Add a cpplint GitHub action
Currently with most checks disabled, except from some runtime checks
that are currently satisfied by the code base.  The idea is to use
this action and gradually enable the checks that we like when these
are fixed, and leave the ones that for whatever reason we do not like
disabled.